[NRRPLT-4023] Make lxml version consistent with CLE.
The CLE and ExDBackend both use lxml to parse SDF files:
- ExDBackend for the rest services to serve world files
- CLE for tests requiring SDF files

Previously, the CLE required lxml 3.4.0 and ExDBackend required
lxml 3.3.3, which caused both to be compiled, installed, and then
uninstalled when the CLE was referenced as a requirement by
ExDBackend.

This patch just bumps the version of the frontend up to 3.4.0 as
there are no incompatibilities or API changes (but are several
bug fixed between 3.3.3 and 3.4.0).

Change-Id: I8bb33deb599dbaf4994701c6186af31753d4c021