[NRRPLT-4188] Don't automatically start state machines after reset.
Removed the code that was automatically starting all state machines after
a reset even though the simulation would be in a paused state. This makes
the logic consistent so that all state machines are halted after a
pause/stop/reset event and the only way to start them is via the start
button.

In the future we may want to explore a mechanism to force the state machines
into a paused state rather than terminating and restarting after each normal
pause event.

Change-Id: I91501bbfc40c68a9ff2b7d475076d7dca1bef058