[NRRPLT-4078][NRRPLT-4249] Let simulation server stop working if it detects that a simulation did not work out correctly; Add Gazebo watchdog
This implementation relies on a server-side script to regularly check that the service create_new_simulation is reachable.
If this is not the case, this script will automatically restart the simulation server.

How to test:
1. Run a simulation.
2. In a different terminal, perform ps -aux | grep gzserver to find gzserver and kill it. This simulates that Gazebo just crashed

Before this change(s), the simulation server would have stopped working entirely. Now, the simulation server will exit the simulation,
publishing a critical error message on the error topic (not yet interpreted by the frontend).
Further, the simulation server marks itself unusable, so it will reject any new simulation requests and will wait for the backend
script to restart the server. This strategy may be a bit too harsh.

Change-Id: Id7a38ad50ff4c7b499a5d5405d72571c0a641e3c