[NRRPLT-4250] Use deployed gazebo models on vizcluster.
This patch modified the Lugano Gazebo launcher to use models that
are deployed to the shared project space via Jenkins rather that copying
them over before starting a simulation.

The previous behavior:
- slowed the launch of an experiment as 500 Mb was deployed each launch
- introduced race conditions where multiple concurrent experiments could
  cause the deletion of the temporary directory linked by ~/.gazebo/models,
  causing running experiments to fail to find assets if another user joined
  or if a user inserted an objected
- another deployment race condition where staging and dev deployments could
  compete for the single ~/.gazebo/models dir and cause strange behavior

Models are deployed to either a dev or staging folder based on their Jenkins
release. The cleserver code checks its own deployment version/type to pick
the correct models to use.

This patch also removes any IO to/from the vizcluster node, except for ssh
calls and the loading of modules from those nodes themselves.

These changes are fully backwards compatible with the faulty behavior of other
releases and should not impact existing deployments.

Change-Id: Ia3208acfb5fb710a7ced0a2af561aa8d45dc7012