[hotfix] Disable Gazebo online model searching.
This patch effectively disables Gazebo's default online model
search for missing models - in this case, the default ground plane
model that is loaded when gzserver is launched. By disabling this
online search and only relying on installed NRP models, the launch
is significantly faster.

When coupled with the patch for [NRRPLT-4250], this means that
dev serve launches go from the order of minutes to a few seconds
(depending on the time for node allocation and the internet connection
for model download speed). This is much more comparable to a local
install.

Change-Id: I8080b5717e77e80b921fcde547335c1d8626574a