[Hotfix] Let CLELauncher know its components earlier.
If something in the experiment goes wrong, then currently the CLELauncher
has no chance to notify its components that have been initialized successfully,
because it only knows them if everything went well. This is a very bad behavior
and leads to errors that for example a cluster allocation is not stopped, if the
brain is not initialized correctly.

Furthermore, there is a small issue with the BIBI functions: The target property of PyNN
has been renamed.

Change-Id: I24b63732d9ba06b1b69f81da14e29adfa4516c3e