[hotfix] Only run one instance of gzserver per vizcluster node.
If a vizcluster node is not heavily used (e.g. the week of a holiday)
and has 2 GPUs and > 8 CPUs available, it is possible for two experiment
launches to end up with the same vizcluster allocation. This causes
the gzserver for the first experiment to be terminated by the launch of
the second.

This is not because multiple gzserver cannot exist on the same server, but
because of how our launch scripts forcefully kill any other gzserver processes.

This patch blacklists any currently allocated nodes so that only one
experiment launch can be run per node.

Change-Id: Id1aeaabeee1ee2383549fa20b3fc2111c83ff8c9