[NRRPLT-4628] Fix for parsing list of allocated nodes.
The previous hotfix failed to sanitize all input before trying to
parse the list of nodes. This means that sometimes the ssh banner info
(a long list of info for users logging in) would be included in the
list to be parsed. This patch clears the buffer before trying to parse
the node list.

Change-Id: I0ad23ef44589551a35949b3e7b3ed29e102eb0f3