[NRRPLT-4604] Better CLE progress bar messages.
This patch adds a few additional status messages that will help in
the future if a simulation launch hangs for a user at a specific state.
There are minor fixes for language consistency and less technical
terminology as well as minor fixes for code structure so that only
the appropriate things are happening in each section.

This was used during profiling (along with additional debug that logged
time in each phase) to determine any particular bad/variable points
during the launch. On average, each state should only be present for
< 3s, the average ism much lower. These messages helped to see that
most of the original "starting gzserver" time was not actually spent
in starting gazebo - it was in the environment configuration and
wait for services that were originally in the same status group.

Change-Id: I4517f6f591d99c228115d4a45226e23746f9ed6c