[NRRPLT-4662] Cleanup ROS nodes/services/topics on shutdown.
This patch ensures we properly cleanup the ROS core between simulation
runs and don't leave any lingering topics or services. Previously,
all of the robot topics and the unreachable Gazebo node would still be
registered with the core.

This patch also includes a small fix to ensure that we always try to
shutdown any launched CLE components after a launch failure.

Change-Id: If78064ba344bf2e9d0b51fe8aa832d2f98f0d8f0