[NRRPLT-4676] Wait for backend rendering/scene on simulation start.
This patch adds a step to startup in CLELauncher to wait for the
backend rendering scene (if needed) to be ready before allowing the
user to interact with the simulation. This is necessary to ensure
that all cameras or other sensors that depend on the rendering
environment are fully initialized before the user can press the
play button, otherwise they will not publish data for a few seconds
until fully loaded.

This is very important for the virtual coach / offline simulations
as everything must be fully initialized before the scripts are
allowed to run.

See the corresponding GazeboRosPackages commit for implementation
details.

Change-Id: I7ef7e7f4873e8f75a037c51d82e25e5f7efe433f