[NRRPLT-4795] Changing a TF with a faulty Brain breaks CLE
Idea behind the change:
Observed behavior: when trying to apply a brain change that make the python script incorrect, then changes are not applied (which is normal), but it breaks the previously working brain. (not normal)
One could chose not to dispose the previous brain until we check that the new brain is OK. However, the logic spans through several layers of abstraction and responsability, and that chnage does not seem straight forward.
Instead, I propose here that we "try setting the brain", and if that high-level instruction fails, we reset the brain with the previously working brain data.

Test -instructions:
 - verify that trying to apply a broken brain does not break the simulation (you can still play it)
 - verify that when brain reset fails, the user gets a relevant message.
     - steps to break the reset: chnage a population name, apply, reset brain

Change-Id: I82ff9b15d476ea54be6a32323a81175a8e4adf63