[NRRPLT-3902][NRRPLT-4844] Move SLURM code to commons and bugfixes.
This patch separates the common functionality from the LuganoVizClusterGazebo
file so that it can be reused by BrainSimulation for distributed nest. This
separated code must be in hbp_nrp_commons to avoid circular references between
the packages. This changes does not add any heavyweight dependencies to commons
and it is still fully standalone.

The functionality is separated as follows:

commons
- SLURM allocation/deallocation
- temporary directory creation, file copying, and cleanup
- environment configuration (from puppet variables)
- ssh sesstions to SLURM master node or any allocated node
- logging / reuse of cle user notifier without any import

cleserver (gazebo specific)
- xvfb server launch
- xvnc launch
- vglconnect method rather than ssh
- gazebo configuration and launch
- watchdog launch

In addition, this patch also resolves [NRRPLT-4844] by providing a generic
remote file copy mechanism that works as intended.

All test coverage has been split to reflect the change as appropriate.

Change-Id: I8c2b273828bb65870a46961f012440046d99c421