[NRRPLT-4845] Ensure watchdog ROS node is always terminated.
The watchdog ROS node is not always properly shutdown on the cluster,
meaning it leaves behind a queue of messages and is not properly
terminated by rosnode cleanup. Reducing the queue to size 1 rather than
10 that can buffer should fix this, but also adding an explicit command
to kill the /Watchdog rosnode ensures that it is guaranteed to be
shutdown and cleaned when the CLE is shutdown.

Change-Id: Icbd227ac3555142d9357fce410ded9b2761853b3