[NRRPLT-4885] Log to stdout for standalone CLE (distributed launch).
This patch redirects logging to stdout when the CLE is launched
standalone and not from ROSCLESimulationFactory.py (e.g. in the case
of a distributed simulation launch).

This is required because the first call to rospy.init_node(...) will
redirect all logging output to ROS loggers rather than to the default
logging config. Since the CLE is launched standalone, it needs to
initialize a dummy node and redirect logging.

Change-Id: I85f59fa77b236ec9029477cbc891a945284cc4af