[NRRPLT-4916] Ensure ROS plugins/nodes always use sim time.
This patch sets the /use_sim_time parameter to true on the ROS
parameter server when the platform is first initialized. This is
required to ensure that all ROS plugins/nodes use sim time properly
rather than real time.

If this parameter is not set properly, the first simulation launch
will not properly use sim time when ros::Time::now() is called,
resulting in joint data that uses real time in messages (e.g. joint
data).

Change-Id: Ia1bfaf7670b622f685a961ba02c92b261e9f5e92