[NRRPLT-5091] Support BIBI Defined Simulation Timestep
This patch adds support for a user defined simulation timestep
to override the default of 20ms. The value is read from the BIBI
if present or a default of 20ms is used for backwards comaptibility.

This value specifies how long to run the robot/brain simulations
before data integration and transfer function execution. Acceptable
range is specified in the BIBI schema definition.

Change-Id: I871f51005dd6fe895c2ff9ea86bb5f16c853cb0b