[hotfix] Fix CLE/Retina Test Regression.
The CLE was updated to check if a specified Retina configuration
file exists, but the corresponding tests were not updated in
ExDBackend and fail. This is not an issue on the build servers yet
as they used the official/released version specified in the
requirements.txt, but local installs use the latest code.

Change-Id: I0698b24744481321a77e3aecda6ded32c735b78a