[NRRPLT-5134] Support roslaunch file specified in ExDConf.
This patch adds roslaunch support to the platform prior to the launch
of Gazebo. The initial launch file should be located in the Experiments
folder, but additional launch files can be referenced in other locations
as long as they are properly in the ROS path or within GazeboRosPackages.

Only local installs are currently supported as the external ROS nodes
will likely not be installed in the deployed environment for custom
user experiments.

Change-Id: I51fd95a82c189ae126b7e660790ea1e8d72558c0