[NRRPLT-5284] Support setting experiment RNG seed.
This patch adds support for setting the Nest and Gazebo RNG
seed. The seed can be specified in the ExDConf, if not present a new
random RNG seed will be generated at each launch.

Documentation will be updated under [NRRPLT-5319] since the ExDConf
documentation is a bit out of date and missing most options.

Change-Id: I22b7bcb5a1add5edb322392b5ad198542769bcad