[NRRPLT-5301] Added POST /transfer-function rest call
This patch adds a new POST call to create new transfer functions
instead of using the PUT call to edit and add tfs.

The previous set_transfer_function function is replaced by two
new functions: add_transfer_function and edit_transfer_function.
When editing a tf, the name of the tf has to be passed as a parameter,
but this is not required anymore when adding a tf, since the name is
extraced from the source code.

A new check on name duplication has been added for whenever a tf is
being edited or added. In case a duplication occurs, the REST call
returns a 403 error.

Change-Id: Ia0e01aee175dd0ade0b7dcd2d7023b355c31ba68