[NRRPLT-5392] Shutdown CLE before Gazebo, extract notificator.
This patch extracts the ROS backend->frontend notification interface
(responsible for publishing state/errors/task status) from the ROSCLEServer
into a higher level class so that the CLE can be shut down before
Gazebo while still notifying the frontend. This allows for a much cleaner
and graceful shutdown flow.

This is required to avoid many (valid) ROS warnings/errors in the
backend log when we killed Gazebo/the notification interface first
and ROS has to deal with ungraceful shutdown. This was seen
intermittently in the logs for the Virtual Coach as ROS (seemingly)
arbitrarily decides when terminal to logs to.

Change-Id: I50597cb20c7e4559dfdd3cce591c3d625059bde3