[NRRPLT-5509] Launch Gazebo recording playback with first file.
This patch changes the playback mechanism to directly launch Gazebo
with the first playback file in the path. Previously, this was all
configured directly in the playback plugins; however, this leads to
issues with Gazebo being slightly misconfigured/confused at a level
that can only be corrected by launching Gazebo in playback mode.

This commit will be updated in the future when the Python storage
server API is available, which will support deployed and local environments
properly and can be made more robust.

Change-Id: I88a42f8337fa434650e32719e8f74ca2267720f7