[NRRPLT-5420] Propagate TF delete failure to REST interface.
This patch propagates the result of a TF request properly from the
TF framework through the various ROS interfaces to the REST client.

This appears to be a legacy issue where the TF deletion/validation of
name to be deleted was not done synchronously and could not be directly
called and resolved. Invalid names sent to delete will no longer
return success.

Change-Id: I8e4b7e23fc969f60fe521c59e889ac22d93b9765