[NRRPLT-4978] Cloning using the storage server
This commit includes changes done to the backend, frontend, and storage server (backend proxy)
to change the cloning process to be agnostic to the kind of storage we use.

Now the backend is querying the storage server to list all the user's experiments
based on his credentials, which he has to provide in a welcome page.

Upon cloning, the backend is reading the experiment_configuration.exc of the
experiment to clone, flattens all the experiment related files into one folder,
and then persistently saves the cloned folder to the storage, and adds an
entry to the database/collab. As a result, a
new cloned experiment is added to the list of available user experiments.

Change-Id: I6b61425c7b6412504a5d1285a3f90c6ad5742597