[NRRPLT-4818] - Load models from generic URIs
This commit includes the changes in the backend, CLE, GazeboRosPackages
in order to make the backend aware of different places on which to
get the models from. Now the user can specify if he wants to load
his models from the template ones (he does not need to change anything)
or from his storage. If he wants from the storage he has to prepend the
model in the .exc or.bibi file with storage://
Then the backend will copy the model from the storage space to a temp
folder on simulation launch.
Change-Id: I8256c3b02c40992bf5152f5030a0a81fc507a1bc