[No Story] Improve backend errors
It happened to Hossain that he had a problem accessing simulation models.
Unfortunately, in that case, Python raises an IOError (as expected) but uses
a MaxRetryError object as the message while we expect that error messages are strings.
This patch solves this problem in order to make sure that better (namely the actual)
error messages are provided in such cases.

Change-Id: If24d08cffc1d01d5497a6edd4f6766aca7bc68f7